cargo clippy --all-targets

+++ src/leetcode/p14_longest_common_prefix/mod.rs
@@ -3,7 +3,7 @@ pub struct Solution {}
 impl Solution {
     pub fn longest_common_prefix(strs: Vec<String>) -> String {
         let strs: Vec<Vec<char>> = strs.iter().map(|x| x.chars().collect()).collect();
-        let mut i = 0 as usize;
+        let mut i = 0usize;
         let num = strs.len() as i32;
         if num == 0 {
             "".to_string()
@@ -11,7 +11,7 @@ impl Solution {
             strs[0].iter().collect()
         } else {
             'a: loop {
-                let mut j = 0 as usize;
+                let mut j = 0usize;
                 while (j as i32) < num - 1 {
                     if !(i < strs[j].len() && i < strs[j + 1].len() && strs[j][i] == strs[j + 1][i])
                     {
